Introduce non-accelerated motion event vectors
For certain applications (such as FPS games) it is necessary to use
non-accelerated motion events (the motion vector that is passed to the
acceleration filter) to get a more natural feeling. Supply this
information by passing both accelerated and non-accelerated motion
vectors to the existing motion event.

Note that the non-accelerated motion event is not equivalent to 'raw'
events as read from devices.

Signed-off-by: Jonas Ådahl <jadahl@gmail.com>